Create good turn
Good Turn
Chef and Chefina are playing with dice. In one turn, both of them roll their dice at once.

They consider a turn to be good if the sum of the numbers on their dice is greater than 
6
6.
Given that in a particular turn Chef and Chefina got 
𝑋
X and 
𝑌
Y on their respective dice, find whether the turn was good.